Cut the postgres trace source, collapse Azure and Bedrock, slim pickers and dead helpers
Wave A of round 2: delete the live-database ingest path (transport-only
duplicate of the file loaders, justification in the PR body), subclass
AzureClient from OpenAICompatibleClient, merge bedrock_converse into
bedrock, drop dead discovery methods and four stranded helpers, and
remove the cookbook page whose opening command fails on the released
CLI.
