+++ docs/README.md
@@ -8,7 +8,6 @@ results, and plans do not live here.
 | File | Purpose |
 |---|---|
 | `usage.md` | Locked CLI map for build, bounded optimize model, optimize router, run, and config. |
-| `cookbook/router.md` | Provider-free local artifact and loopback runtime walkthrough. |
 | `reference/providers.md` | Catalog providers, first-build `--provider` flags, environment variables, Azure endpoint and deployment rules, and the Bedrock credential chain. |
 | `reference/ingest.md` | Current declared local trace source contract for every supported source. |
 | `reference/immutable-real-trace-rag.md` | Immutable real-trace retrieval provenance, leakage, persistence, and historical restoration contract. |
