+++ wmo/cli/build_cmd.py
@@ -51,10 +51,7 @@ _PROJECT_ARGUMENT = typer.Argument(..., metavar="PROJECT", help="Local project n
 _TRACE_FILE_ARGUMENT = typer.Argument(
     ...,
     metavar="TRACES",
-    help=(
-        "Local trace export in the declared --source format, or the JSON source declaration "
-        "of a postgres table."
-    ),
+    help="Local trace export in the declared --source format.",
 )
 _PROVIDER_OPTION = typer.Option(
     None,
@@ -103,7 +100,7 @@ def build(
     Args:
         project: Safe local project identifier below ``<root>/projects``.
         trace_file: Explicit local canonical trace export.
-        source: Declared local-export format, or ``postgres`` for a table declaration.
+        source: Declared local-export format.
         root: Local ``.wmo`` artifact root.
         world_model: Optional configured alias override for this project.
         judge: Optional configured alias override for this project.
