+++ wmo/cli/model_picker_test.py
@@ -27,7 +27,6 @@ from wmo.common.models import (
     ModelRecord,
     PricingSource,
     ProviderConnection,
-    ProviderModelSelection,
     SetupRole,
     resolve_discovered_model,
 )
@@ -198,8 +197,8 @@ def test_the_setup_written_from_confirmed_answers_carries_verified_metadata() ->
         chosen,
         roles=roles,
         endpoints=(_endpoint(),),
-        existing_connections=(),
-        existing_models=(),
+        known_existing_connections=(),
+        known_existing_aliases=(),
     )
 
     setup = result.setup
@@ -226,10 +225,6 @@ def test_already_configured_models_are_not_written_again() -> None:
         pricing_source=PricingSource.CONFIGURED,
         configured=True,
     )
-    existing_models = (
-        ProviderModelSelection(alias="luna", connection="openai", model="gpt-5.6-luna"),
-    )
-
     result = build_result(
         (configured, _EMBEDDER),
         roles=assign_roles(
@@ -239,8 +234,8 @@ def test_already_configured_models_are_not_written_again() -> None:
         )
         or pytest.fail("roles were not assigned"),
         endpoints=(_endpoint(),),
-        existing_connections=(_OPENAI,),
-        existing_models=existing_models,
+        known_existing_connections=("openai",),
+        known_existing_aliases=("luna",),
     )
 
     assert [model.alias for model in result.setup.models] == ["embedder"]
@@ -371,8 +366,8 @@ def test_summary_states_providers_models_roles_prices_and_credential_behavior() 
         chosen,
         roles=roles,
         endpoints=(_endpoint(),),
-        existing_connections=(),
-        existing_models=(),
+        known_existing_connections=(),
+        known_existing_aliases=(),
     )
     render_summary(result, chosen=chosen, endpoints=(_endpoint(),), console=console)
 
@@ -400,8 +395,8 @@ def test_summary_names_the_aws_credential_chain_for_bedrock() -> None:
         chosen,
         roles=roles,
         endpoints=(_endpoint(), _endpoint(bedrock)),
-        existing_connections=(),
-        existing_models=(),
+        known_existing_connections=(),
+        known_existing_aliases=(),
     )
     render_summary(
         result,
@@ -428,8 +423,8 @@ def test_summary_names_confirmed_router_candidates() -> None:
         chosen,
         roles=roles,
         endpoints=(_endpoint(),),
-        existing_connections=(),
-        existing_models=(),
+        known_existing_connections=(),
+        known_existing_aliases=(),
     )
     render_summary(result, chosen=chosen, endpoints=(_endpoint(),), console=console)
 
@@ -450,8 +445,8 @@ def test_a_connection_only_used_by_unselected_models_is_not_written() -> None:
         chosen,
         roles=roles,
         endpoints=(_endpoint(), _endpoint(other)),
-        existing_connections=(),
-        existing_models=(),
+        known_existing_connections=(),
+        known_existing_aliases=(),
     )
 
     assert [connection.name for connection in result.setup.connections] == ["openai"]
