+++ wmo/common/judging/__init__.py
@@ -32,7 +32,6 @@ from wmo.common.judging.lm import (
     LMJudge,
     RawDimensionJudgment,
     RawJudgment,
-    judge_response_schema,
 )
 from wmo.common.judging.prompts import PromptDefinition
 from wmo.common.judging.proposal import (
@@ -97,5 +96,4 @@ __all__ = [
     "write_router_lineage_split",
     "write_rubric_proposal_evidence",
     "verify_persisted_calibration",
-    "judge_response_schema",
 ]
