+++ wmo/common/tasks/__init__.py
@@ -1,6 +1,6 @@
 """Canonical representative task contracts and immutable artifact reads."""
 
-from wmo.common.tasks.store import LoadedTaskSet, load_task_set, resolve_task_set
+from wmo.common.tasks.store import LoadedTaskSet, load_task_set
 from wmo.common.tasks.task import TaskCase, TaskSet, ToolSchema
 
 __all__ = [
@@ -9,5 +9,4 @@ __all__ = [
     "TaskSet",
     "ToolSchema",
     "load_task_set",
-    "resolve_task_set",
 ]
